+++ src/lib/swapper/swappers/CowSwapper/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -8,13 +8,13 @@ import type { CowSwapQuoteResponse } from 'lib/swapper/swappers/CowSwapper/types
 import {
   COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS,
   COW_SWAP_VAULT_RELAYER_ADDRESS,
-  getFullAppData,
   ORDER_KIND_SELL,
 } from 'lib/swapper/swappers/CowSwapper/utils/constants'
 import { cowService } from 'lib/swapper/swappers/CowSwapper/utils/cowService'
 import {
   assertValidTrade,
   getCowswapNetwork,
+  getFullAppData,
   getNowPlusThirtyMinutesTimestamp,
   getSupportedChainIds,
   getValuesFromQuoteResponse,
